bsr(api) move file to context target-profile-management
