+++ api/lib/domain/usecases/index.js
@@ -406,7 +406,7 @@ const createOrganizationPlacesLot = require('./create-organization-places-lot.js
 const createOrganizationsWithTagsAndTargetProfiles = require('./create-organizations-with-tags-and-target-profiles.js');
 const createPasswordResetDemand = require('./create-password-reset-demand.js');
 const createSession = require('./create-session.js');
-const createStage = require('./create-stage.js');
+const createStage = require('./target-profile-management/create-stage.js');
 const createTag = require('./create-tag.js');
 const createTargetProfile = require('./create-target-profile.js');
 const createTraining = require('./create-training.js');
@@ -622,7 +622,7 @@ const updateOrganizationInformation = require('./update-organization-information
 const updateOrganizationLearnerDependentUserPassword = require('./update-organization-learner-dependent-user-password.js');
 const updateParticipantExternalId = require('./update-participant-external-id.js');
 const updateSession = require('./update-session.js');
-const updateStage = require('./update-stage.js');
+const updateStage = require('./target-profile-management/update-stage.js');
 const updateStudentNumber = require('./update-student-number.js');
 const updateTargetProfile = require('./update-target-profile.js');
 const updateTraining = require('./update-training.js');
