+++ api/lib/infrastructure/repositories/target-profile-management/stage-collection-repository.js
@@ -16,4 +16,8 @@ module.exports = {
     const rawIds = await knex('stages').insert(stageCollection.stages).onConflict('id').merge().returning('id');
     return rawIds.map((rawId) => rawId.id);
   },
+
+  delete(stageCollection) {
+    return knex('stage').where('id', stageCollection.id).delete();
+  },
 };
