+++ api/tests/unit/domain/usecases/target-profile-management/create-stage_test.js
@@ -1,6 +1,6 @@
 const { expect, domainBuilder, catchErrSync } = require('../../../test-helper');
-const usecases = require('../../../../lib/domain/usecases/index.js');
-const { InvalidStageError } = require('../../../../lib/domain/errors');
+const usecases = require('../../../../../lib/domain/usecases/index.js');
+const { InvalidStageError } = require('../../../../../lib/domain/errors');
 
 describe('Unit | UseCases | create-stage', function () {
   let stageCollection;
